fix: изменить createStore на config, перенести константы для работы с хранилищем в файл store
