+++ src/index.js
@@ -3,20 +3,9 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from "./components/app/app";
 import reportWebVitals from './reportWebVitals';
-
 import { Provider } from 'react-redux';
-import { createStore, applyMiddleware, compose } from 'redux';
-import {thunk} from 'redux-thunk';
 
-import { rootReducer } from './services/reducers/index';
-
-
-const composeEnhancers =
-  window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose
-
-const enhancer = composeEnhancers(applyMiddleware(thunk));
-
-const store = createStore(rootReducer, enhancer);
+import { store } from './services/store';
 
 
 const root = ReactDOM.createRoot(
